move parsing to ownfile
