+++ src/iniconfig/exceptions.py
@@ -1,9 +1,9 @@
-
 from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from typing_extensions import Final
+
 
 class ParseError(Exception):
     path: Final[str]
